# api/docs/swagger.yaml
definitions:
  handler.Response:
    properties:
      data: {}
      description:
        type: string
      error: {}
      status:
        type: integer
    type: object
  models.AdminCreate:
    properties:
      addres:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone_number:
        type: string
    type: object
  models.AdminUpdate:
    properties:
      addres:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone_number:
        type: string
    type: object
  models.Banner:
    properties:
      banner_image:
        type: string
      created_at:
        type: string
      delete_at:
        type: string
      id:
        type: string
      updated_at:
        type: string
    type: object
  models.BannerCreate:
    properties:
      banner_image:
        type: string
    type: object
  models.BannerUpdate:
    properties:
      banner_image:
        type: string
      id:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /e_commerce/api/v1/admin:
    get:
      consumes:
      - application/json
      description: Get List Admin
      operationId: get_list_admin
      parameters:
      - description: offset
        in: query
        name: offset
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get List Admin
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Create Admin
      operationId: create_admins
      parameters:
      - description: CreateAdminRequest
        in: body
        name: Admin
        required: true
        schema:
          $ref: '#/definitions/models.AdminCreate'
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Admin
      tags:
      - Admin
  /e_commerce/api/v1/admin/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Admin
      operationId: delete_admin
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Admin
      tags:
      - Admin
    get:
      consumes:
      - application/json
      description: Get By ID Admin
      operationId: get_by_id_admin
      parameters:
      - description: id
        in: path
        name: id
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get By ID Admin
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Update Admin
      operationId: update_admin
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateAdminRequest
        in: body
        name: Admin
        required: true
        schema:
          $ref: '#/definitions/models.AdminUpdate'
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Admin
      tags:
      - Admin
  /e_commerce/api/v1/banner:
    get:
      consumes:
      - application/json
      description: Get List Banner
      operationId: get_list_banner
      parameters:
      - description: offset
        in: query
        name: offset
        type: string
      - description: limit
        in: query
        name: limit
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get List Banner
      tags:
      - Banner
    post:
      consumes:
      - application/json
      description: Create Banner
      operationId: create_banner
      parameters:
      - description: CreateBannerRequest
        in: body
        name: Banner
        required: true
        schema:
          $ref: '#/definitions/models.BannerCreate'
      responses:
        "200":
          description: Success Request
          schema:
            $ref: '#/definitions/models.Banner'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Banner
      tags:
      - Banner
  /e_commerce/api/v1/banner/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Banner
      operationId: delete_banner
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Banner
      tags:
      - Banner
    get:
      consumes:
      - application/json
      description: Get By ID Banner
      operationId: get_by_id_banner
      parameters:
      - description: id
        in: path
        name: id
        type: string
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get By ID Banner
      tags:
      - Banner
    put:
      consumes:
      - application/json
      description: Update Banner
      operationId: update_banner
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: UpdateBannerRequest
        in: body
        name: Banner
        required: true
        schema:
          $ref: '#/definitions/models.BannerUpdate'
      responses:
        "200":
          description: Success Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Banner
      tags:
      - Banner
...
swagger: "2.0"